Модуляция. Вторая проверка
