+++ js/main.js
@@ -1,2 +1,2 @@
 // eslint-disable-next-line no-unused-vars
-import {receivedImages} from './data.js';
+import { images } from './data.js';
