+++ js/utils/check-string-max-length.js
@@ -1,0 +1,8 @@
+const checkStringMaxLength = (stringCheck, maxLength) => {
+  if (stringCheck.length > maxLength) {
+    return false;
+  }
+  return true;
+};
+
+export { checkStringMaxLength };
