+++ js/utils/get-random-integer.js
@@ -1,4 +1,3 @@
-// за основу взят код отсюда https://developer.mozilla.org/ru/docs/Web/JavaScript/Reference/Global_Objects/Math/random
 const getRandomInteger = (min, max) => {
   let minimumInteger = Math.ceil(min);
   let maximumInteger = Math.floor(max);
@@ -14,4 +13,4 @@ const getRandomInteger = (min, max) => {
   return Math.floor(Math.random() * (maximumInteger - minimumInteger + 1)) + minimumInteger;
 };
 
-export {getRandomInteger};
+export { getRandomInteger };
